+++ Changelog.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## April 7th, 2024
+### **[HouseRules | Experimental]**
+- Modifiers button can now be pressed while in Continue tab
+- Added 8 more new modifiers
+
+## April 7th, 2024
 *(at literally 12am, again)*
 ### **[HouseRules]**
 - Added